backend/requirements.txt
# === DEPENDENCIAS CORE DJANGO ===
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-cors-headers>=4.3
django-environ>=0.11
python-dotenv>=1.0

# === BASE DE DATOS ===
psycopg2-binary>=2.9
dj-database-url>=2.1.0

# === UTILIDADES BÁSICAS ===
requests>=2.31.0
pillow>=10.0.0

# ===== SISTEMA RAG =====

# === LANGCHAIN ECOSYSTEM ===
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1

# === GOOGLE DRIVE API ===
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0
google-auth>=2.0.0

# === PROCESAMIENTO DE DOCUMENTOS ===
pypdf>=4.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pymupdf>=1.23.0
python-docx>=1.1.0
python-pptx>=0.6.0
openpyxl>=3.1.0

# === BASE DE DATOS VECTORIAL ===
chromadb>=0.4.0

# === EMBEDDINGS Y NLP (CPU-only, sin CUDA) ===
# IMPORTANTE: Instalar torch CPU PRIMERO para evitar versión CUDA
torch==2.3.1+cpu
torchvision==0.18.1+cpu
torchaudio==2.3.1+cpu
-f https://download.pytorch.org/whl/torch_stable.html
sentence-transformers>=2.2.0
tiktoken>=0.5.0
transformers>=4.35.0

# === ANÁLISIS Y UTILIDADES ===
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0
python-magic>=0.4.27

# === EVALUACIÓN Y MÉTRICAS ===
datasets>=2.18.0
evaluate>=0.4.0
ragas>=0.1.10

# === GOOGLE GEMINI API ===
google-generativeai>=0.3.0

# === LOGGING Y MONITOREO ===
python-json-logger>=2.0.0

# === PRODUCCIÓN ===
gunicorn>=21.2.0
whitenoise>=6.6.0

# === DEPENDENCIAS ADICIONALES DETECTADAS ===
langchain-openai>=0.0.1
langchain-groq>=0.0.1
markdown>=3.4.0
python-multipart>=0.0.6
mysql-connector-python>=8.0.33
docx>=0.2.4
